Refactor:
Move the navigation bar from the Classroom feature component to it's own component.

Add the drop shadow to the nav bar as per Figma.
