+++ src/components/NavigationBar.scss
@@ -1,0 +1,3 @@
+.navbar {
+    box-shadow: 0 2px 4px rgba($black, .15)  !important;
+}
