+++ src/index.scss
@@ -10,6 +10,7 @@
 
 @import './components/classroomheader.scss';
 @import './components/coursecard.scss';
+@import './components/NavigationBar.scss';
 @import './features/classroom/classroom.scss';
 
 body {
